+++ bitframe-authentication/services/ktor/src/commonMain/kotlin/bitframe/authentication/signin/SignInServiceKtor.kt
@@ -5,7 +5,10 @@ import bitframe.authentication.spaces.Space
 import bitframe.response.response.decodeResponseFromString
 import bitframe.service.MiniService
 import bitframe.service.config.KtorClientConfiguration
+import bitframe.service.utils.JsonContent
+import io.ktor.client.features.*
 import io.ktor.client.request.*
+import io.ktor.client.statement.*
 import io.ktor.content.*
 import io.ktor.http.*
 import kotlinx.serialization.json.Json
@@ -21,13 +24,12 @@ class SignInServiceKtor @JvmOverloads constructor(
     private val scope = config.scope
     override fun signIn(credentials: SignInCredentials): Later<LoginConundrum> = scope.later {
         validate(credentials)
-        val json = http.post<String>(path) {
-            body = TextContent(
-                text = Json.encodeToString(SignInCredentials.serializer(), credentials),
-                contentType = ContentType.Application.Json
-            )
+        val resp = try {
+            http.post(path) { body = JsonContent(credentials) }
+        } catch (exp: ClientRequestException) {
+            exp.response
         }
-        Json.decodeResponseFromString(LoginConundrum.serializer(), json).response()
+        Json.decodeResponseFromString(LoginConundrum.serializer(), resp.readText()).response()
     }
 
     override fun resolve(space: Space): Later<Session.SignedIn> = scope.later {