+++ pi-monitor/pi-monitor-client/browser/react/src/jsMain/kotlin/pimonitor/index.kt
@@ -7,21 +7,18 @@ import kotlinext.js.jso
 import kotlinx.browser.document
 import kotlinx.extensions.By
 import kotlinx.extensions.get
-import logging.ConsoleAppender
-import logging.Logging
 import org.w3c.dom.HTMLDivElement
 import pimonitor.authentication.signup.SignUp
-import pimonitor.authentication.signup.legacy.SignUpLegacy
 import pimonitor.evaluation.business.BusinessContainer
 import reakt.setContent
 
 fun main() = document.get<HTMLDivElement>(By.id("root")).setContent {
     val client = client(jso {
         appId = "test-client"
-        simulationTime = 3000
+//        url = "http://localhost:8080"
+        simulationTime = 1500
     })
     val version: String by konfig()
-    Logging.init(ConsoleAppender())
     Bitframe(
         client = client,
         routeRenderers = mapOf(