+++ pi-monitor/pi-monitor-client/sdks/full/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpIntent.kt
@@ -2,14 +2,12 @@
 
 package pimonitor.authentication.signup
 
-import pimonitor.authentication.signup.legacy.IndividualFormFields
-import pimonitor.authentication.signup.legacy.OrganisationFormFields
 import pimonitor.monitors.SignUpParams
 import kotlin.js.JsExport
 
 sealed class SignUpIntent {
-    data class SelectRegisterAsIndividual(val fields: IndividualFormFields?) : SignUpIntent()
-    data class SelectRegisterAsBusiness(val fields: OrganisationFormFields?) : SignUpIntent()
+    object SelectRegisterAsIndividual : SignUpIntent()
+    object SelectRegisterAsBusiness : SignUpIntent()
 
     sealed class Submit(open val params: SignUpParams) : SignUpIntent() {
         data class IndividualForm(override val params: SignUpParams.Individual) : Submit(params)