+++ pi-monitor/pi-monitor-services/core/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpService.kt
@@ -2,10 +2,8 @@
 
 package pimonitor.authentication.signup
 
-import bitframe.authentication.signin.LoginConundrum
 import contacts.Email
 import later.Later
-import pimonitor.Monitor
 import pimonitor.monitors.SignUpParams
 import kotlin.js.JsExport
 
@@ -32,8 +30,6 @@ abstract class SignUpService {
             }
         }
     }
-
-    abstract fun registerIndividuallyAs(person: IndividualRegistrationParams): Later<LoginConundrum>
-    abstract fun register(business: Monitor.Business, representedBy: Monitor.Person): Later<LoginConundrum>
+    
     abstract fun signUp(params: SignUpParams): Later<SignUpResult>
 }