+++ pi-monitor/pi-monitor-services/ktor/src/commonMain/kotlin/pimonitor/authentication/signup/SignUpServiceKtor.kt
@@ -1,15 +1,15 @@
 package pimonitor.authentication.signup
 
-import bitframe.authentication.signin.LoginConundrum
 import bitframe.response.response.decodeResponseFromString
 import bitframe.service.MiniService
 import bitframe.service.config.KtorClientConfiguration
 import bitframe.service.utils.JsonContent
+import io.ktor.client.features.*
 import io.ktor.client.request.*
+import io.ktor.client.statement.*
 import kotlinx.serialization.json.Json
 import later.Later
 import later.later
-import pimonitor.Monitor
 import pimonitor.monitors.SignUpParams
 
 class SignUpServiceKtor(
@@ -17,21 +17,14 @@ class SignUpServiceKtor(
 ) : SignUpService(), MiniService {
     private val client = config.http
     private val scope = config.scope
-    override fun registerIndividuallyAs(
-        person: IndividualRegistrationParams
-    ): Later<LoginConundrum> = scope.later {
-        val json = client.post<String>(config.url + "/api/authentication/sign-up") {
-            body = JsonContent(person)
+
+    override fun signUp(params: SignUpParams): Later<SignUpResult> = scope.later {
+        validate(params)
+        val resp = try {
+            client.post(config.url + "/api/authentication/sign-up") { body = JsonContent(params) }
+        } catch (err: ClientRequestException) {
+            err.response
         }
-        val res = Json.decodeResponseFromString(LoginConundrum.serializer(), json)
-        res.response()
-    }
-
-    override fun register(business: Monitor.Business, representedBy: Monitor.Person): Later<LoginConundrum> {
-        TODO("Not yet implemented")
-    }
-
-    override fun signUp(params: SignUpParams): Later<SignUpResult> {
-        TODO("Not yet implemented")
+        Json.decodeResponseFromString(SignUpResult.serializer(), resp.readText()).response()
     }
 }