+++ pi-monitor/pi-monitor-services/ktor/src/commonMain/kotlin/pimonitor/evaluation/businesses/BusinessServiceKtor.kt
@@ -2,12 +2,12 @@ package pimonitor.evaluation.businesses
 
 import bitframe.service.config.KtorClientConfiguration
 import later.Later
-import pimonitor.Monitor
+import pimonitor.monitored.MonitoredBusiness
 
 class BusinessServiceKtor(
     private val config: KtorClientConfiguration
 ) : BusinessService() {
-    override fun all(): Later<List<Monitor.Business>> {
+    override fun all(): Later<List<MonitoredBusiness>> {
         TODO()
     }
 }